# contracts/fixtures/openapi/zone-v2.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nexus Zone v1 API",
    "version": "0.1.0"
  },
  "paths": {
    "/v2/zone-capabilities": {
      "get": {
        "responses": {
          "200": {
            "description": "OK"
          },
          "default": {
            "description": "Contract error shape (code/message/retryable)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "properties": {
                    "detail": {
                      "type": "object",
                      "required": [
                        "code",
                        "retryable"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "retryable": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v2/zones": {
      "get": {
        "responses": {
          "200": {
            "description": "OK"
          },
          "default": {
            "description": "Contract error shape (code/message/retryable)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "properties": {
                    "detail": {
                      "type": "object",
                      "required": [
                        "code",
                        "retryable"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "retryable": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "responses": {
          "202": {
            "description": "Accepted"
          },
          "default": {
            "description": "Contract error shape (code/message/retryable)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "properties": {
                    "detail": {
                      "type": "object",
                      "required": [
                        "code",
                        "retryable"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "retryable": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v2/zones/{zone_id}": {
      "get": {
        "responses": {
          "200": {
            "description": "OK"
          },
          "default": {
            "description": "Contract error shape (code/message/retryable)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "properties": {
                    "detail": {
                      "type": "object",
                      "required": [
                        "code",
                        "retryable"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "retryable": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "patch": {
        "responses": {
          "202": {
            "description": "Accepted"
          },
          "default": {
            "description": "Contract error shape (code/message/retryable)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "properties": {
                    "detail": {
                      "type": "object",
                      "required": [
                        "code",
                        "retryable"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "retryable": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "responses": {
          "202": {
            "description": "Accepted"
          },
          "default": {
            "description": "Contract error shape (code/message/retryable)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "properties": {
                    "detail": {
                      "type": "object",
                      "required": [
                        "code",
                        "retryable"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "retryable": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v2/zone-operations/{operation_id}": {
      "get": {
        "responses": {
          "200": {
            "description": "OK"
          },
          "default": {
            "description": "Contract error shape (code/message/retryable)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "properties": {
                    "detail": {
                      "type": "object",
                      "required": [
                        "code",
                        "retryable"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "retryable": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v2/zones/{zone_id}/grants": {
      "get": {
        "responses": {
          "200": {
            "description": "OK"
          },
          "default": {
            "description": "Contract error shape (code/message/retryable)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "properties": {
                    "detail": {
                      "type": "object",
                      "required": [
                        "code",
                        "retryable"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "retryable": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "responses": {
          "202": {
            "description": "Accepted"
          },
          "default": {
            "description": "Contract error shape (code/message/retryable)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "properties": {
                    "detail": {
                      "type": "object",
                      "required": [
                        "code",
                        "retryable"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "retryable": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v2/zones/{zone_id}/grants/{grant_id}": {
      "get": {
        "responses": {
          "200": {
            "description": "OK"
          },
          "default": {
            "description": "Contract error shape (code/message/retryable)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "properties": {
                    "detail": {
                      "type": "object",
                      "required": [
                        "code",
                        "retryable"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "retryable": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "responses": {
          "202": {
            "description": "Accepted"
          },
          "default": {
            "description": "Contract error shape (code/message/retryable)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "properties": {
                    "detail": {
                      "type": "object",
                      "required": [
                        "code",
                        "retryable"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "retryable": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v2/auth/zone-delegations": {
      "post": {
        "responses": {
          "202": {
            "description": "Accepted"
          },
          "default": {
            "description": "Contract error shape (code/message/retryable)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "properties": {
                    "detail": {
                      "type": "object",
                      "required": [
                        "code",
                        "retryable"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "retryable": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "request_schema": "contracts/auth/v1/zone-delegation-issue-request.schema.json",
        "response_schema": "contracts/auth/v1/zone-delegation.schema.json"
      }
    },
    "/v2/auth/zone-delegations/{delegation_id}": {
      "get": {
        "responses": {
          "200": {
            "description": "OK"
          },
          "default": {
            "description": "Contract error shape (code/message/retryable)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "properties": {
                    "detail": {
                      "type": "object",
                      "required": [
                        "code",
                        "retryable"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "retryable": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "response_schema": "contracts/auth/v1/zone-delegation.schema.json"
      },
      "delete": {
        "responses": {
          "202": {
            "description": "Accepted"
          },
          "default": {
            "description": "Contract error shape (code/message/retryable)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "detail"
                  ],
                  "properties": {
                    "detail": {
                      "type": "object",
                      "required": [
                        "code",
                        "retryable"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string"
                        },
                        "retryable": {
                          "type": "boolean"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}